managed/managed-config-error.https.html

<!DOCTYPE html>
<title>Managed Configuration: Unsupported values</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

promise_test(t => {
  return promise_rejects_js(t, TypeError, navigator.managed.getManagedConfiguration(-1));
}, "Number instead of keys");

promise_test(t => {
return promise_rejects_js(t, TypeError, navigator.managed.getManagedConfiguration());
}, "Empty key list");

promise_test(t => {
return promise_rejects_js(t, TypeError, navigator.managed.getManagedConfiguration({"a" : 2}));
}, "Dictionary instead of list");

</script>
